inha win8 metrowin contribution / "porting mode" programming - many IDEs in different platforms and OSes... m_build contribution... Eu te amo Thommy, Thomas Borregaard Sørensen!!
Mummi contribution!!!! Feliz 61 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 7 Ultimate 64-bit Service Pack 1 - with Portuguese - Brazil - Language for non-Unicode programs - ~10Mpbs NET Virtua Internet at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!

-----8nnnnnnnn8
++++++++++

+++ os/metrowin_application.cpp
@@ -714,7 +714,7 @@ namespace metrowin
    #endif
    }
 
-   unsigned long application::get_thread_id()
+   uint32_t application::get_thread_id()
    {
       return ::GetCurrentThreadId();
    }
